Fix fp16 overflow in UNIT, MUNIT, DCLGAN, MIUDiff, UVCGAN under AMP
- unit.py: clamp logvar to [-30, 10] in reparameterize and kl_loss to prevent exp() overflow
- munit.py: clamp AdaIN gamma/beta to [-10, 10] in AdaINBottleneck to prevent unbounded style scaling
- base_models.py: cast q/k to float32 in info_nce before matmul to avoid fp16 overflow at low temperature
- miudiff.py: fp32 softmax in AttentionBlock to prevent attention score overflow
- uvcgan.py: cast to float32 around MultiheadAttention in TransformerBlock for the same reason

+++ base_models.py
@@ -399,8 +399,8 @@ def denorm01(x: torch.Tensor) -> torch.Tensor:
 
 def info_nce(q: torch.Tensor, k: torch.Tensor, temperature: float = 0.07) -> torch.Tensor:
     """InfoNCE loss with one-to-one positives by index."""
-    q = F.normalize(q, dim=1)
-    k = F.normalize(k, dim=1)
+    q = F.normalize(q.float(), dim=1)
+    k = F.normalize(k.float(), dim=1)
     logits = q @ k.t() / temperature
     labels = torch.arange(q.size(0), device=q.device)
     return F.cross_entropy(logits, labels)
